+++ src/pages/Greeter/Greeter.tsx
@@ -18,7 +18,7 @@ export const GreeterPage: FunctionComponent = () => {
       </div>
       <p>
         Select your settings, and this app will guide ya&apos; thru a jack-off session. You can adjust most of the aspects of the game, as
-        well as select some porn to help with &quot;motivation&quot;. It&apos;s all pulled from <a href="https://e621.net">E621.net</a>.{' '}
+        well as select some porn to help with &quot;motivation&quot;. It&apos;s all pulled from <a href="https://e621.net">E621.net</a>, <a href="https://www.redgifs.com">redgifs.com</a> or from your own local collection.{' '}
         <strong>You are responsible for what you choose to look at.</strong>
       </p>
       <p>