+++ src/release-notes.md
@@ -19,6 +19,7 @@
 - add separate list of porn specifically for cumming to (anthrofox22)
 - deploy to GitHub pages (https://anthrofox22.github.io/joi.how/) (anthrofox22)
 - add support for loading videos from RedGifs based on a search term and sexuality preferences (anthrofox22)
+- add support for loading files from local sources (anthrofox22)
 
 ## 1.6.0 - Fri, 15 Apr 2022 10:26:26 -0400